requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pymongo==4.6.2
python-dotenv==1.0.1
gql==3.5.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.2.1
plotly==5.19.0 